Refactor pages into separate files
