+++ lib/main.dart
@@ -1,48 +1,11 @@
 import 'package:flutter/material.dart';
 import 'dish_storage.dart';
-import 'package:flutter/services.dart';
 import 'ui_components.dart';
-import 'package:intl/intl.dart';
+import 'models.dart';
+import 'widgets/dish_list.dart';
+import 'pages/weeks_menu_page.dart';
+import 'pages/history_page.dart';
 
-class AppConstants {
-  static const double dishFontSize = 16.0;
-  static const double dishVerticalSpacing = 8.0;
-}
-
-const Color kCardColor = Color(0xFFBBDEFB); // Colors.blue[100]
-
-const Map<DishCategory, Color> kCategoryColors = {
-  DishCategory.egg: Color(0xFFFFF9C4), // light yellow
-  DishCategory.pork: Color(0xFFFFCCBC), // light orange
-  DishCategory.beef: Color(0xFFD7CCC8), // light brown
-  DishCategory.fish: Color(0xFFB3E5FC), // light blue
-  DishCategory.tofu: Color(0xFFC8E6C9), // light green
-  DishCategory.other: Color(0xFFE0E0E0), // light grey
-};
-
-enum DishCategory { egg, pork, beef, fish, tofu, other }
-
-class Dish {
-  final String name;
-  int count;
-  DishCategory category;
-
-  Dish({required this.name, required this.count, required this.category});
-
-  factory Dish.fromJson(Map<String, dynamic> json) {
-    return Dish(
-      name: json['name'] as String,
-      count: json['count'] as int,
-      category: DishCategory.values[json['category'] as int],
-    );
-  }
-
-  Map<String, dynamic> toJson() => {
-        'name': name,
-        'count': count,
-        'category': category.index,
-      };
-}
 
 void main() {
   runApp(const MainApp());
@@ -97,8 +60,8 @@ class _MainAppState extends State<MainApp> {
   late List<Dish> _dishes;
   DishCategory? _selectedCategory;
   bool _loading = true;
-  final GlobalKey<_WeeksMenuPageState> _weeksMenuKey =
-      GlobalKey<_WeeksMenuPageState>();
+  final GlobalKey<WeeksMenuPageState> _weeksMenuKey =
+      GlobalKey<WeeksMenuPageState>();
 
   @override
   void initState() {
@@ -150,7 +113,7 @@ class _MainAppState extends State<MainApp> {
                 ? const Center(child: CircularProgressIndicator())
                 : TabBarView(
                     children: [
-                      WeeksMenuPage(key: _weeksMenuKey),
+                      WeeksMenuPage(key: _weeksMenuKey, allDishes: _dishes),
                       Padding(
                         padding: const EdgeInsets.symmetric(horizontal: 16.0),
                         child: Column(
@@ -260,661 +223,5 @@ class _MainAppState extends State<MainApp> {
   }
 }
 
-class DishList extends StatelessWidget {
-  final List<Dish> dishes;
-  final void Function(String dishName) onIncrement;
 
-  const DishList({super.key, required this.dishes, required this.onIncrement});
 
-  void _openDishDetail(BuildContext context, Dish dish) {
-    Navigator.push(
-      context,
-      MaterialPageRoute(
-        builder: (context) => DishDetailPage(
-          dish: dish,
-          onIncrement: () => onIncrement(dish.name),
-        ),
-      ),
-    );
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    final sortedDishes = [...dishes]..sort((a, b) {
-        final countCompare = b.count.compareTo(a.count);
-        if (countCompare != 0) return countCompare;
-        return a.name.toLowerCase().compareTo(b.name.toLowerCase());
-      });
-
-    return SingleChildScrollView(
-      child: Column(
-        crossAxisAlignment: CrossAxisAlignment.start,
-        children: [
-          SizedBox(height: AppConstants.dishVerticalSpacing),
-          ...sortedDishes.map((dish) => Column(
-                children: [
-                  SizedBox(
-                    width: double.infinity,
-                    child: InkWell(
-                      onTap: () => _openDishDetail(context, dish),
-                      child: Card(
-                        color: kCategoryColors[dish.category] ?? kCardColor,
-                        child: Padding(
-                          padding: const EdgeInsets.all(12.0),
-                          child: Row(
-                            mainAxisAlignment: MainAxisAlignment.spaceBetween,
-                            children: [
-                              Expanded(
-                                child: Text(
-                                  dish.name,
-                                  style: TextStyle(
-                                    fontSize: AppConstants.dishFontSize,
-                                    color: Colors.black87,
-                                    fontWeight: FontWeight.bold,
-                                  ),
-                                  maxLines: 1,
-                                  overflow: TextOverflow.ellipsis,
-                                ),
-                              ),
-                              Text(
-                                '${dish.count}',
-                                style: const TextStyle(
-                                  fontSize: 16,
-                                  color: Colors.black54,
-                                ),
-                              ),
-                              const SizedBox(width: 8),
-                              Text(
-                                categoryToString(dish.category),
-                                style: const TextStyle(
-                                  fontSize: 14,
-                                  color: Colors.black38,
-                                ),
-                              ),
-                            ],
-                          ),
-                        ),
-                      ),
-                    ),
-                  ),
-                  SizedBox(height: AppConstants.dishVerticalSpacing),
-                ],
-              )),
-        ],
-      ),
-    );
-  }
-}
-
-class DishDetailPage extends StatefulWidget {
-  final Dish dish;
-  final VoidCallback onIncrement;
-
-  const DishDetailPage(
-      {super.key, required this.dish, required this.onIncrement});
-
-  @override
-  State<DishDetailPage> createState() => _DishDetailPageState();
-}
-
-class _DishDetailPageState extends State<DishDetailPage> {
-  late DishCategory? _selectedCategory;
-  List<String> _customCategories = [];
-  String? _selectedCustomCategory;
-  bool _loading = true;
-
-  @override
-  void initState() {
-    super.initState();
-    _selectedCategory = widget.dish.category;
-    _loadCustomCategories();
-  }
-
-  Future<void> _loadCustomCategories() async {
-    final cats = await DishStorageCategories.loadCustomCategories();
-    setState(() {
-      _customCategories = cats;
-      _loading = false;
-      if (_selectedCategory == DishCategory.other &&
-          cats.contains(widget.dish.name)) {
-        _selectedCustomCategory = widget.dish.name;
-      }
-    });
-  }
-
-  Future<void> _saveCustomCategories() async {
-    await DishStorageCategories.saveCustomCategories(_customCategories);
-  }
-
-  void _updateCategory(DishCategory? cat, [String? customCat]) async {
-    setState(() {
-      _selectedCategory = cat;
-      _selectedCustomCategory = customCat;
-    });
-    if (cat == DishCategory.other &&
-        customCat != null &&
-        customCat.isNotEmpty) {
-      if (!_customCategories.contains(customCat)) {
-        setState(() {
-          _customCategories.add(customCat);
-        });
-        await _saveCustomCategories();
-      }
-      setState(() {
-        widget.dish.category = DishCategory.other;
-      });
-    } else if (cat != null) {
-      setState(() {
-        widget.dish.category = cat;
-      });
-    }
-    // Save dish update to storage
-    final dishes = await DishStorage.loadDishes(MainApp.dishes);
-    final idx = dishes.indexWhere((d) => d.name == widget.dish.name);
-    if (idx != -1) {
-      dishes[idx].category = widget.dish.category;
-      await DishStorage.saveDishes(dishes);
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (_loading) {
-      return const Center(child: CircularProgressIndicator());
-    }
-    return Scaffold(
-      appBar: AppBar(
-        title: Text(widget.dish.name),
-      ),
-      body: Padding(
-        padding: const EdgeInsets.all(24.0),
-        child: Column(
-          crossAxisAlignment: CrossAxisAlignment.start,
-          children: [
-            const SizedBox(height: 16),
-            Row(
-              children: [
-                const Text('Category: ',
-                    style: TextStyle(fontSize: 16, color: Colors.black54)),
-                DropdownButton<DishCategory>(
-                  value: _selectedCategory,
-                  items: [
-                    ...DishCategory.values.map((cat) => DropdownMenuItem(
-                          value: cat,
-                          child: Text(categoryToString(cat)),
-                        )),
-                  ],
-                  onChanged: (cat) {
-                    if (cat == DishCategory.other) {
-                      setState(() {
-                        _selectedCategory = cat;
-                        _selectedCustomCategory = null;
-                      });
-                    } else {
-                      _updateCategory(cat);
-                    }
-                  },
-                ),
-                if (_selectedCategory == DishCategory.other)
-                  const SizedBox(width: 8),
-                if (_selectedCategory == DishCategory.other)
-                  DropdownButton<String>(
-                    value: _selectedCustomCategory,
-                    hint: const Text('Custom'),
-                    items: [
-                      ..._customCategories.map((cat) => DropdownMenuItem(
-                            value: cat,
-                            child: Text(cat),
-                          )),
-                      const DropdownMenuItem<String>(
-                        value: '__add_new__',
-                        child: Text('Add new...'),
-                      ),
-                    ],
-                    onChanged: (val) async {
-                      if (val == '__add_new__') {
-                        final newCat = await showDialog<String>(
-                          context: context,
-                          builder: (context) {
-                            String input = '';
-                            return AlertDialog(
-                              title: const Text('Add new category'),
-                              content: TextField(
-                                autofocus: true,
-                                decoration: const InputDecoration(
-                                    labelText: 'Category name'),
-                                onChanged: (v) => input = v,
-                              ),
-                              actions: [
-                                TextButton(
-                                  onPressed: () => Navigator.of(context).pop(),
-                                  child: const Text('Cancel'),
-                                ),
-                                ElevatedButton(
-                                  onPressed: () =>
-                                      Navigator.of(context).pop(input),
-                                  child: const Text('Add'),
-                                ),
-                              ],
-                            );
-                          },
-                        );
-                        if (newCat != null && newCat.trim().isNotEmpty) {
-                          _updateCategory(DishCategory.other, newCat.trim());
-                        }
-                      } else if (val != null) {
-                        _updateCategory(DishCategory.other, val);
-                      }
-                    },
-                  ),
-              ],
-            ),
-            const SizedBox(height: 16),
-            const Text(
-              'Placeholder info about this dish.',
-              style: TextStyle(fontSize: 18),
-            ),
-            const SizedBox(height: 24),
-            Text(
-              'Cooked: ${widget.dish.count} times',
-              style: const TextStyle(fontSize: 18, color: Colors.black87),
-            ),
-            const SizedBox(height: 16),
-            ElevatedButton(
-              onPressed: () {
-                widget.onIncrement();
-                Navigator.of(context).pop();
-              },
-              child: const Text('I cooked this!'),
-            ),
-          ],
-        ),
-      ),
-    );
-  }
-}
-
-// Helper for category display names
-String categoryToString(DishCategory category) {
-  switch (category) {
-    case DishCategory.egg:
-      return 'Egg';
-    case DishCategory.pork:
-      return 'Pork';
-    case DishCategory.beef:
-      return 'Beef';
-    case DishCategory.fish:
-      return 'Fish';
-    case DishCategory.tofu:
-      return 'Tofu';
-    case DishCategory.other:
-      return 'Other';
-  }
-}
-
-class WeeksMenuPage extends StatefulWidget {
-  const WeeksMenuPage({super.key});
-
-  static const List<String> daysOfWeek = [
-    'Monday',
-    'Tuesday',
-    'Wednesday',
-    'Thursday',
-    'Friday',
-    'Saturday',
-    'Sunday',
-  ];
-
-  @override
-  State<WeeksMenuPage> createState() => _WeeksMenuPageState();
-}
-
-class _WeeksMenuPageState extends State<WeeksMenuPage> {
-  Map<String, WeeksMenuEntry> _menu = {
-    for (final day in WeeksMenuPage.daysOfWeek) day: WeeksMenuEntry()
-  };
-  String _label = '';
-  bool _loading = true;
-
-  @override
-  void initState() {
-    super.initState();
-    _loadMenuAndLabel();
-  }
-
-  Future<void> _loadMenuAndLabel() async {
-    final loaded =
-        await DishStorageWeeksMenu.loadWeeksMenu(WeeksMenuPage.daysOfWeek);
-    String label = await DishStorageWeeksMenu.loadWeeksMenuLabel();
-    if (label.isEmpty) {
-      // Default to current week (ISO week number)
-      final now = DateTime.now();
-      final weekOfYear = int.parse(DateFormat('w').format(now));
-      final year = now.year;
-      label = 'Week $weekOfYear, $year';
-    }
-    setState(() {
-      _menu = loaded;
-      _label = label;
-      _loading = false;
-    });
-  }
-
-  Future<void> _saveMenu() async {
-    await DishStorageWeeksMenu.saveWeeksMenu(_menu);
-  }
-
-  Future<void> _saveLabel(String label) async {
-    await DishStorageWeeksMenu.saveWeeksMenuLabel(label);
-  }
-
-  Future<void> _selectDish(BuildContext context, String day) async {
-    final allDishes = MainApp.dishes;
-    DishCategory? filterCategory;
-    String? selected = _menu[day]?.dishName;
-    final result = await showDialog<String>(
-      context: context,
-      builder: (context) {
-        return StatefulBuilder(
-          builder: (context, setState) {
-            final filteredDishes = filterCategory == null
-                ? allDishes
-                : allDishes.where((d) => d.category == filterCategory).toList();
-            return AlertDialog(
-              title: Text('Select dish for $day'),
-              content: SizedBox(
-                width: double.maxFinite,
-                child: Column(
-                  mainAxisSize: MainAxisSize.min,
-                  children: [
-                    Row(
-                      children: [
-                        const Text('Filter: '),
-                        DropdownButton<DishCategory?>(
-                          value: filterCategory,
-                          hint: const Text('All'),
-                          items: [
-                            const DropdownMenuItem<DishCategory?>(
-                              value: null,
-                              child: Text('All'),
-                            ),
-                            ...DishCategory.values
-                                .map((cat) => DropdownMenuItem<DishCategory?>(
-                                      value: cat,
-                                      child: Text(categoryToString(cat)),
-                                    )),
-                          ],
-                          onChanged: (cat) =>
-                              setState(() => filterCategory = cat),
-                        ),
-                      ],
-                    ),
-                    const SizedBox(height: 12),
-                    Expanded(
-                      child: GridView.count(
-                        crossAxisCount: 1,
-                        childAspectRatio: 3.5,
-                        shrinkWrap: true,
-                        children: filteredDishes.map((dish) {
-                          final isSelected = dish.name == selected;
-                          return GestureDetector(
-                            onTap: () => Navigator.of(context).pop(dish.name),
-                            child: Card(
-                              color: isSelected
-                                  ? Theme.of(context)
-                                      .colorScheme
-                                      .primary
-                                      .withOpacity(0.2)
-                                  : Colors.white,
-                              shape: RoundedRectangleBorder(
-                                side: isSelected
-                                    ? BorderSide(
-                                        color: Theme.of(context)
-                                            .colorScheme
-                                            .primary,
-                                        width: 2)
-                                    : BorderSide(
-                                        color: Colors.grey.shade300, width: 1),
-                                borderRadius: BorderRadius.circular(8),
-                              ),
-                              child: Padding(
-                                padding: const EdgeInsets.symmetric(
-                                    horizontal: 16.0, vertical: 8.0),
-                                child: Row(
-                                  children: [
-                                    Expanded(
-                                      child: Text(
-                                        dish.name,
-                                        style: TextStyle(
-                                          fontWeight: isSelected
-                                              ? FontWeight.bold
-                                              : FontWeight.normal,
-                                          color: isSelected
-                                              ? Theme.of(context)
-                                                  .colorScheme
-                                                  .primary
-                                              : Colors.black87,
-                                        ),
-                                        overflow: TextOverflow.ellipsis,
-                                      ),
-                                    ),
-                                    Container(
-                                      padding: const EdgeInsets.symmetric(
-                                          horizontal: 8, vertical: 2),
-                                      decoration: BoxDecoration(
-                                        color: kCategoryColors[dish.category]
-                                            ?.withOpacity(0.5),
-                                        borderRadius: BorderRadius.circular(6),
-                                      ),
-                                      child: Text(
-                                        categoryToString(dish.category),
-                                        style: const TextStyle(fontSize: 12),
-                                      ),
-                                    ),
-                                  ],
-                                ),
-                              ),
-                            ),
-                          );
-                        }).toList(),
-                      ),
-                    ),
-                  ],
-                ),
-              ),
-              actions: [
-                TextButton(
-                  onPressed: () => Navigator.of(context).pop(),
-                  child: const Text('Cancel'),
-                ),
-              ],
-            );
-          },
-        );
-      },
-    );
-    if (result != null) {
-      setState(() {
-        _menu[day] = WeeksMenuEntry(dishName: result, cooked: false);
-      });
-      _saveMenu();
-    }
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (_loading) {
-      return const Center(child: CircularProgressIndicator());
-    }
-    return ListView.separated(
-      padding: const EdgeInsets.all(16),
-      itemCount: WeeksMenuPage.daysOfWeek.length + 1,
-      separatorBuilder: (context, index) => const SizedBox(height: 12),
-      itemBuilder: (context, index) {
-        if (index == 0) {
-          // Label at the top
-          return Row(
-            children: [
-              Expanded(
-                child: TextFormField(
-                  initialValue: _label,
-                  decoration: const InputDecoration(
-                    labelText: 'Week label',
-                    border: OutlineInputBorder(),
-                  ),
-                  onChanged: (val) {
-                    setState(() {
-                      _label = val;
-                    });
-                    _saveLabel(val);
-                  },
-                ),
-              ),
-            ],
-          );
-        }
-        final day = WeeksMenuPage.daysOfWeek[index - 1];
-        final entry = _menu[day] ?? WeeksMenuEntry();
-        final dish = entry.dishName;
-        final cooked = entry.cooked;
-        Color cardColor = kCardColor;
-        if (dish != null) {
-          final selectedDish = MainApp.dishes.firstWhere(
-            (d) => d.name == dish,
-            orElse: () =>
-                Dish(name: '', count: 0, category: DishCategory.other),
-          );
-          cardColor = kCategoryColors[selectedDish.category] ?? kCardColor;
-        }
-        return WeekMenuCard(
-          day: day,
-          dish: dish,
-          cooked: cooked,
-          cardColor: cardColor,
-          onTap: () => _selectDish(context, day),
-          onCookedChanged: (val) {
-            if (dish == null) return;
-            setState(() {
-              _menu[day] = WeeksMenuEntry(dishName: dish, cooked: val ?? false);
-            });
-            _saveMenu();
-          },
-        );
-      },
-    );
-  }
-
-  Future<void> clearWeek() async {
-    // Save to history first
-    final label = _label;
-    final menuCopy = {
-      for (final e in _menu.entries)
-        e.key:
-            WeeksMenuEntry(dishName: e.value.dishName, cooked: e.value.cooked)
-    };
-    final entry = WeeksMenuHistoryEntry(label: label, menu: menuCopy);
-    final history = await DishStorageWeeksMenuHistory.loadWeeksMenuHistory();
-    history.add(entry);
-    await DishStorageWeeksMenuHistory.saveWeeksMenuHistory(history);
-    // Clear current week
-    setState(() {
-      _menu = {
-        for (final day in WeeksMenuPage.daysOfWeek) day: WeeksMenuEntry()
-      };
-    });
-    DishStorageWeeksMenu.saveWeeksMenu(_menu);
-  }
-}
-
-class HistoryPage extends StatefulWidget {
-  const HistoryPage({super.key});
-
-  @override
-  State<HistoryPage> createState() => _HistoryPageState();
-}
-
-class _HistoryPageState extends State<HistoryPage> {
-  List<WeeksMenuHistoryEntry> _history = [];
-  bool _loading = true;
-
-  @override
-  void initState() {
-    super.initState();
-    _loadHistory();
-  }
-
-  Future<void> _loadHistory() async {
-    final history = await DishStorageWeeksMenuHistory.loadWeeksMenuHistory();
-    setState(() {
-      _history = history.reversed.toList(); // Show most recent first
-      _loading = false;
-    });
-  }
-
-  @override
-  Widget build(BuildContext context) {
-    if (_loading) {
-      return const Center(child: CircularProgressIndicator());
-    }
-    if (_history.isEmpty) {
-      return const Center(child: Text('No previous weeks yet.'));
-    }
-    return ListView.separated(
-      padding: const EdgeInsets.all(16),
-      itemCount: _history.length,
-      separatorBuilder: (context, index) => const SizedBox(height: 16),
-      itemBuilder: (context, index) {
-        final entry = _history[index];
-        return Card(
-          child: Padding(
-            padding: const EdgeInsets.all(16.0),
-            child: Column(
-              crossAxisAlignment: CrossAxisAlignment.start,
-              children: [
-                Text(entry.label,
-                    style: const TextStyle(
-                        fontSize: 18, fontWeight: FontWeight.bold)),
-                const SizedBox(height: 8),
-                ...WeeksMenuPage.daysOfWeek.map((day) {
-                  final dishEntry = entry.menu[day];
-                  return Row(
-                    children: [
-                      SizedBox(
-                        width: 100,
-                        child: Text(day,
-                            style:
-                                const TextStyle(fontWeight: FontWeight.bold)),
-                      ),
-                      if (dishEntry?.dishName != null)
-                        Expanded(
-                          child: Text(
-                            dishEntry!.dishName!,
-                            style: TextStyle(
-                              color: dishEntry.cooked
-                                  ? Colors.green
-                                  : Colors.black87,
-                              decoration: dishEntry.cooked
-                                  ? TextDecoration.lineThrough
-                                  : null,
-                            ),
-                          ),
-                        )
-                      else
-                        const Expanded(child: Text('-')),
-                      if (dishEntry?.cooked == true)
-                        const Padding(
-                          padding: EdgeInsets.only(left: 8.0),
-                          child:
-                              Icon(Icons.check, color: Colors.green, size: 18),
-                        ),
-                    ],
-                  );
-                }).toList(),
-              ],
-            ),
-          ),
-        );
-      },
-    );
-  }
-}
